ms_ner/requirements.txt
fastAPI
uvicorn
python-dotenv
nltk==3.7
redis
celery